Extract duplicated `FeedLoaderStub` into a separate file
